+++ src/connection.cpp
@@ -2,7 +2,7 @@
 
 using namespace pneatm;
 
-Connection::Connection (const unsigned int innovId, const unsigned int inNodeId, const unsigned int outNodeId, const unsigned int inNodeRecu, float weight, bool enabled):
+Connection::Connection (const unsigned int innovId, const unsigned int inNodeId, const unsigned int outNodeId, const unsigned int inNodeRecu, double weight, bool enabled):
     innovId (innovId),
     inNodeId (inNodeId),
     outNodeId (outNodeId),
